Добавляет задание 5 модуля 'Модуляция'
