+++ js/util.js
@@ -1,0 +1,41 @@
+
+//рандомное число
+
+const getRandomInteger = (min, max) => {
+  const lower = Math.ceil(Math.min(Math.abs(min), Math.abs(max)));
+  const upper = Math.floor(Math.max(Math.abs(min), Math.abs(max)));
+  const result = Math.random() * (upper - lower + 1) + lower;
+  return Math.floor(result);
+};
+
+//рандомное число без повтора
+
+const createRandomIdFromRangeGenerator = (min, max) => {
+  const previousValues = [];
+
+  return function () {
+    let currentValue = getRandomInteger(min, max);
+    if (previousValues.length >= (max - min + 1)) {
+      console.error('Перебраны все числа из диапазона от ' + min + ' до ' + max);
+      return null;
+    }
+    while (previousValues.includes(currentValue)) {
+      currentValue = getRandomInteger(min, max);
+    }
+    previousValues.push(currentValue);
+    return currentValue;
+  };
+};
+
+const createIdGenerator = (min, max) => {
+  min = Math.ceil(min);
+  max = Math.floor(max);
+  return Math.floor(Math.random() * (max - min + 1) + min);
+};
+
+const generateId = createRandomIdFromRangeGenerator(1, 25);
+const generatePhotoId = createRandomIdFromRangeGenerator(1, 25);
+const generateLikes = createRandomIdFromRangeGenerator(15, 200);
+const generateCommentatorId = createRandomIdFromRangeGenerator(1, 500);
+
+export { generateId, generatePhotoId, generateLikes, generateCommentatorId, createIdGenerator,getRandomInteger };
